[Font] Harden bitmap font loading and metrics
* Propagate font selection, refresh and setup failures instead of masking them

* Validate bitmap font offsets and buffer bounds before cache transformations

* Fix font metrics, wrapping and shared cache concurrency edge cases

+++ src/font/font_structs.h
@@ -4,8 +4,8 @@
 struct FontCharacter {
    int16_t  Width;
    int16_t  Advance;
-   uint16_t Offset;
-   uint16_t OutlineOffset;
+   uint32_t Offset;
+   uint32_t OutlineOffset;
 };
 
 typedef const std::lock_guard<std::recursive_mutex> CACHE_LOCK;
